// src/api/generated/models/createArea.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Rukun API
 * Standard API contract for Core R.
 * OpenAPI spec version: 1.0.0
 */
import type { CreateAreaKind } from './createAreaKind';
import type { CreateAreaParentId } from './createAreaParentId';

export interface CreateArea {
  kind: CreateAreaKind;
  /** Required RW UUID for an RT; forbidden for RW */
  parent_id?: CreateAreaParentId;
  /** @maxLength 20 */
  code: string;
  /** @maxLength 100 */
  name: string;
}